+++ scripts/settings.js
@@ -81,6 +81,7 @@ export const TVA_CONFIG = {
     syncImages: false,
     overrideCategory: false,
     category: 'Token',
+    missingImages: [{ document: 'all', image: CONST.DEFAULT_TOKEN }],
   },
   permissions: {
     popups: {